refactor(application): use kernel `NotificationMessage` and move tracking related services in a per se package
